Remove config that stopped tab completion in the command buffer
I was not getting any tab completion after pressing ':', turns out it
was this piece of config that was causing it. I'm not exactly sure what
this config does as I just copied form the README. If I suddenly miss
what it was doing, I'll look into this some more.